refactor: Amount 클래스 제거 및 Formatter 추가
- getter와 toString을 가지는 기존의 Amount 클래스를 제거
  - 돈의 수량은 int로 사용
  - 돈의 포맷을 담당하는 메소드를 가지는 MoneyFormatter 추가 (outputView 패키지)
    - UI의 책임을 분리하기 위함
