+++ src/main/java/christmas/controller/PromotionController.java
@@ -4,7 +4,6 @@ import static christmas.controller.ErrorMessages.INVALID_DATE_INPUT;
 import static christmas.controller.ErrorMessages.INVALID_ORDER_INPUT;
 
 import christmas.domain.badge.Badge;
-import christmas.domain.amount.Amount;
 import christmas.domain.event.EventResult;
 import christmas.domain.order.Order;
 import christmas.domain.visitDate.VisitDate;
@@ -31,8 +30,8 @@ public class PromotionController {
 
         outputView.displayPreviewMessage(visitDate.getDate());
 
-        Amount entirePrice = new Amount(order.getEntirePrice());
-        outputView.displayEntirePrice(entirePrice.toString());
+        int entirePrice = order.getEntirePrice();
+        outputView.displayEntirePrice(entirePrice);
 
         EventResult eventResult = EventResult.ofOrderAndDate(order, visitDate.getDate());
         outputView.displayGifts(eventResult.getGiftsNameAndCount());
@@ -43,7 +42,7 @@ public class PromotionController {
         int totalDiscountEventDiscountAmount = eventResult.getTotalDiscountEventDiscountAmount();
         int totalGiftEventDiscountAmount = eventResult.getTotalGiftEventDiscountAmount();
         outputView.displayTotalDiscountAmount(totalDiscountEventDiscountAmount, totalGiftEventDiscountAmount);
-        outputView.displayPaymentAmount(entirePrice.amount(), totalDiscountEventDiscountAmount);
+        outputView.displayPaymentAmount(entirePrice, totalDiscountEventDiscountAmount);
 
         int totalBenefitAmount = totalDiscountEventDiscountAmount + totalGiftEventDiscountAmount;
         Badge badge = Badge.fromBenefitAmount(totalBenefitAmount);
